+++ src/kodo_js/decoder.hpp
@@ -20,28 +20,6 @@
 namespace kodo_js
 {
 
-    // template<class Decoder>
-    // uint32_t decoder_recode(Decoder& decoder)
-    // {
-    //     std::vector<uint8_t> payload(decoder.payload_size());
-    //     return decoder.recode(payload.data());
-    // }
-
-    // template<class Decoder>
-    // void decoder_decode(Decoder& decoder, const std::string& payload)
-    // {
-    //     decoder.decode((uint8_t*)payload.c_str());
-    // }
-
-    // template<class Decoder>
-    // void decoder_decode_symbol(Decoder& decoder, const std::string& symbol_data,
-    //     const std::string& symbol_coefficients)
-    // {
-    //     decoder.decode_symbol(
-    //         (uint8_t*)symbol_data.c_str(),
-    //         (uint8_t*)symbol_coefficients.c_str());
-    // }
-
     template<class Decoder>
     bool decoder_is_complete(Decoder& decoder)
     {
@@ -70,18 +48,6 @@ namespace kodo_js
         return decoder.is_symbol_uncoded(index);
     }
 
-    // template<class Decoder>
-    // bool decoder_has_partial_decoding_tracker(Decoder& decoder)
-    // {
-    //     return decoder.has_partial_decoding_tracker();
-    // }
-
-    // template<class Decoder>
-    // bool decoder_is_partial_complete(Decoder& decoder)
-    // {
-    //     return decoder.is_partial_complete();
-    // }
-
     template<class Decoder>
     void decoder_read_payload(Decoder& decoder, const std::string payload)
     {
@@ -94,37 +60,18 @@ namespace kodo_js
         return decoder.symbols_seen();
     }
 
-    template<class Decoder>
-    std::string decoder_write_payload(Decoder& decoder)
-    {
-        std::vector<uint8_t> payload(decoder.payload_size());
-        decoder.write_payload(payload.data());
-        return std::string(payload.begin(), payload.end());
-    }
-
     template<template<class, class> class Coder, class Field>
     void decoder(const std::string& name)
     {
-        // typedef Coder<Field, TraceTag> decoder_type;
-
-        // coder<Coder, Field, TraceTag>(std::string("decoder") + name)
         typedef Coder<Field, meta::typelist<>> decoder_type;
 
         coder<Coder, Field>(std::string("decoder") + name)
-            // .function("recode", &decoder_recode<decoder_type>)
-            // .function("decode", &decoder_decode<decoder_type>)
-            // .function("decode_symbol", &decoder_decode_symbol<decoder_type>)
             .function("is_complete", &decoder_is_complete<decoder_type>)
             .function("symbols_uncoded", &decoder_symbols_uncoded<decoder_type>)
-            .function("copy_symbols", &decoder_copy_from_symbols<decoder_type>)
+            .function("copy_from_symbols", &decoder_copy_from_symbols<decoder_type>)
             .function("is_symbol_uncoded", &decoder_is_symbol_uncoded<decoder_type>)
             .function("read_payload", &decoder_read_payload<decoder_type>)
             .function("symbols_seen", &decoder_symbols_seen<decoder_type>)
-            // .function("is_partial_complete", &decoder_is_partial_complete<decoder_type>)
-            // .function("has_partial_decoding_tracker",
-            //           &decoder_has_partial_decoding_tracker<decoder_type>)
-            .function("write_feedback", &decoder_write_payload<decoder_type>)
-
         ;
     }
 }