+++ NEWS.rst
@@ -6,4 +6,6 @@ every change, see the Git log.
 
 Latest
 ------
-* tbd
+* Major: Updated Kodo-Js bindings, to depend upon Kod 30.0.0 instead of Kodo 19
+* Major: Renamed `decode` to `read_payload` and `encode` and `recode` to `write_payload`
+* Major: Renamed `copy_symbols` to `copy_from_symbols`